+++ src/main/java/gift/authentication/infrastructure/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package gift.authentication.infrastructure;
 
-import gift.authentication.TokenProvider;
+import gift.authentication.service.TokenProvider;
 import gift.core.domain.authentication.Token;
 import gift.core.domain.authentication.exception.AuthenticationExpiredException;
 import gift.core.domain.authentication.exception.AuthenticationFailedException;
